+++ tests/int_bitbucket.rs
@@ -159,6 +159,7 @@ fn test_bitbucket_remote_scan() -> Result<()> {
         no_ignore_if_contains: false,
         validation_retries: 1,
         validation_timeout: 10,
+        full_validation_response: false,
     };
 
     let global_args = GlobalArgs {