+++ tests/int_dedup.rs
@@ -179,6 +179,7 @@ rules:
         no_ignore_if_contains: false,
         validation_retries: 1,
         validation_timeout: 10,
+        full_validation_response: false,
     };
 
     let global_args = GlobalArgs {