+++ tests/int_vulnerable_files.rs
@@ -165,6 +165,7 @@ impl TestContext {
             no_ignore_if_contains: false,
             validation_retries: 1,
             validation_timeout: 10,
+            full_validation_response: false,
         };
 
         let loaded = RuleLoader::from_rule_specifiers(&scan_args.rules)
@@ -302,6 +303,7 @@ impl TestContext {
             no_ignore_if_contains: false,
             validation_retries: 1,
             validation_timeout: 10,
+            full_validation_response: false,
         };
 
         let global_args = GlobalArgs {